Add a new machine learner to do one-nearest-neighbor
This is a dumber and simpler machine learner for testing. When
making a decision, the 'closest' feature set to the test set
is found in the training set, and the configuration for that
feature set in the training data is used for the test feature
set.

ChangeLog:

	* CMakeLists.txt: Build the 1NN machine learner.
	* include/mageec/ML/1NN.h: Added file.
	* lib/Driver.cpp (main): Initialize the 1NN machine learner.
	* lib/ML/1NN.cpp: Added file.
	* lib/Util.cpp (Util::read64LE): Fix to actually read 64 bits of
	data.

+++ lib/Util.cpp
@@ -65,7 +65,7 @@ void write16LE(std::vector<uint8_t> &buf, unsigned value) {
 }
 
 uint64_t read64LE(std::vector<uint8_t>::const_iterator &it) {
-  unsigned res = 0;
+  uint64_t res = 0;
   res |= static_cast<uint64_t>(*it);
   res |= static_cast<uint64_t>(*(it + 1)) << 8;
   res |= static_cast<uint64_t>(*(it + 2)) << 16;
